+++ tests/print_line_length.c
@@ -4,7 +4,7 @@
 // clang-format off
 //< -f 1
 //> Options:
-//>   -b, --bool                   Option with a description so long that it must be split to the second line (default: false)
+//>   -b, --bool                   Option with a description so long that it must be split to the second line
 //>       --option-with-long-name  A long option (default: 12)
 //>   -f, --float                  A float option
 //>       --str                    A string option (default: "some default")