+++ tests/print_no_min_desc_length.c
@@ -14,7 +14,6 @@
 //>                                to the
 //>                                second
 //>                                line
-//>                                (default: false)
 //>       --option-with-long-name  A long
 //>                                option
 //>                                (default: 12)