+++ tests/print_padding.c
@@ -5,7 +5,6 @@
 //> Options:
 //>   -b, --bool                           Option with a description so long that it
 //>                                        must be split to the second line
-//>                                        (default: false)
 //>       --option-with-long-name          A long option (default: 12)
 //>   -f, --float                          A float option
 //>       --str                            A string option (default: "some default")